luge imu : new jostle detection : log the standard deviation instead of its square

+++ libDCM/rmat.c
@@ -219,13 +219,13 @@ union longww accum32 ;
 extern int32_t omegagyro32X[] ;
 extern union longww theta_32[];
 
-int64_t _gyro_sum_of_squares = 0 ;
-int16_t _total_samples = 0 ;
+uint64_t _gyro_sum_of_squares = 0 ;
+uint16_t _total_samples = 0 ;
 int32_t _gyro_sum[] = { 0 , 0 , 0 };
-int64_t gyro_sum_of_squares = 0 ;
-int16_t total_samples = 0 ;
+uint64_t gyro_sum_of_squares = 0 ;
+uint16_t total_samples = 0 ;
 int32_t gyro_sum[] = { 0 , 0 , 0 };
-int64_t stdev_sqr = 0 ;
+uint64_t stdev_sqr = 0 ;
 
 extern int16_t check_for_jostle ;
 uint16_t jostle_counter = 0 ;
@@ -242,9 +242,9 @@ static inline void read_gyros(void)
     _gyro_sum[1] += (int32_t) omegagyro[1];
     _gyro_sum[2] += (int32_t) omegagyro[2];
     
-    _gyro_sum_of_squares += (int64_t)__builtin_mulss( omegagyro[0],omegagyro[0])
-            + (int64_t)__builtin_mulss( omegagyro[1],omegagyro[1])
-            + (int64_t)__builtin_mulss( omegagyro[2],omegagyro[2]) ;
+    _gyro_sum_of_squares += (uint64_t)__builtin_mulss( omegagyro[0],omegagyro[0])
+            + (uint64_t)__builtin_mulss( omegagyro[1],omegagyro[1])
+            + (uint64_t)__builtin_mulss( omegagyro[2],omegagyro[2]) ;
     
     if ((udb_heartbeat_counter % HEARTBEAT_HZ )== 0) jostle_counter ++ ;
 			if ( jostle_counter == JOSTLE_CHECK_PERIOD ) 
@@ -257,7 +257,7 @@ static inline void read_gyros(void)
         total_samples = _total_samples ;
         _total_samples = 0 ;
         // compute the average of the sum of the squares of the gyro signals
-        gyro_sum_of_squares = _gyro_sum_of_squares / ((int64_t)total_samples);
+        gyro_sum_of_squares = _gyro_sum_of_squares / ((uint64_t)total_samples);
         _gyro_sum_of_squares = 0 ;
         // compute the average of the gyro signals
         gyro_sum[0] = _gyro_sum[0]/((int32_t)total_samples) ;
@@ -270,10 +270,10 @@ static inline void read_gyros(void)
         // compute the variance, which is the mean of the squares of the samples
         // minus the products of the means of the samples, using the classic equation
         
-        stdev_sqr = gyro_sum_of_squares 
+        stdev_sqr = (uint64_t)(gyro_sum_of_squares 
                 - ((int64_t)gyro_sum[0])*((int64_t)gyro_sum[0])
                 - ((int64_t)gyro_sum[1])*((int64_t)gyro_sum[1])
-                - ((int64_t)gyro_sum[2])*((int64_t)gyro_sum[2]) ;
+                - ((int64_t)gyro_sum[2])*((int64_t)gyro_sum[2])) ;
         
         if (stdev_sqr > GYRO_VARIANCE_MARGIN)
         {
